moved music button on lobby
